+++ protected/humhub/docs/CHANGELOG.md
@@ -73,6 +73,7 @@ HumHub Change Log
 - Enh: Directory view templates cleanups
 - Fix: All LDAP Users have been disabled and not reenabled by hourly cronjob if ldap server not reachable.
 - Enh: Cleanup authentication layout file 
+- Fix: Console image converter memory limit allocation
 
 1.2.0-beta.2 (February 24, 2017)
 --------------------------------